Script to build and pack to npx package

+++ scripts/remove-vk-dbs.sh
@@ -3,6 +3,9 @@
 PREFFIX="auto_devs_vk"
 db_list=$(psql -d postgres -c "SELECT datname FROM pg_database WHERE datname LIKE '$PREFFIX%'")
 for db in $db_list; do
-    echo "Removing database: $db"
-    psql -d postgres -c "SELECT pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = '$db' AND pid <> pg_backend_pid();" && dropdb $db
+    # do not process if db is not starts with PREFFIX
+    if [[ "$db" == $PREFFIX* ]]; then
+        echo "Removing database: $db"
+        psql -d postgres -c "SELECT pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = '$db' AND pid <> pg_backend_pid();" && dropdb $db
+    fi
 done
